ingest/requirements.txt
slack-sdk==3.30.0
langchain==0.1.20
google-api-python-client==2.134.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
sentence-transformers==3.0.1
GitPython==3.1.43
pymilvus==2.4.4
llama-index-core==0.10.48
llama-index-readers-web==0.1.19
llama-index-readers-github==0.1.9
llama-index-vector-stores-milvus==0.1.20
llama-index-embeddings-langchain==0.1.2
trafilatura==1.10.0
nbconvert==7.16.4
ipython==8.25.0
upstash-redis==1.1.0
jupyter-notebook-parser==0.1.4
youtube-transcript-api==0.6.3
